Implicitly convert strings (paths), regex, and boolean to RouteMatcher.
This significantly reduces the overloads of get, post, put, and delete.
Also, it is more flexible: an arbitrary number of route matchers may now
be strung together.  The route matches if all matchers return Some, and
the parameters are the sum of all the routes.  A route matcher will not
run if any route matcher before it fails.